requirements.txt
Pillow>=9.0.0
pystray>=0.19.0